posts to'liq CRUD va Faqat joylagan odam postn off

+++ src/controllers/post.contr.ts
@@ -29,32 +29,23 @@ class PostController {
     async getPosts(req: Request, res: Response) {
         try {
             let token: any = req.headers.token;
-            if (!token) {
-                return res.status(401).json({
-                    error: 'Token not found'
-                });
-            }
             const id = JWT.VERIFY(token).id;
             if (!id) {
                 return res.status(401).json({
                     error: 'unknown'
                 });
             }
-            const user: IUser | null = await UserSchema.findById(id).populate("posts");
-            res.json(user);
+            let posts: IPost[] | null = await Post.find({ user: id });
+            res.json(posts);
         } catch (error) {
             res.status(500).json({ error: 'Postlarni olishda xatolik yuz berdi' });
         }
     }
+
     // Postni olish
     async getPostById(req: Request, res: Response) {
         try {
             let token: any = req.headers.token;
-            if (!token) {
-                return res.status(401).json({
-                    error: 'Token not found'
-                });
-            }
             const id = JWT.VERIFY(token).id;
             if (!id) {
                 return res.status(401).json({
@@ -75,12 +66,20 @@ class PostController {
     // Postni yangilash
     async updatePost(req: Request, res: Response) {
         try {
+            let token: any = req.headers.token;
+            const id = JWT.VERIFY(token).id;
+            if (!id) {
+                return res.status(401).json({
+                    error: 'unknown'
+                });
+            }
             const { title, content } = req.body;
-            const post: IPost | null = await Post.findByIdAndUpdate(req.params.id, { title, content }, { new: true });
-            if (post) {
+            let post: IPost | null = await Post.findById(req.params.id);
+            if (post && post.user == id) {
+                let post: IPost | null = await Post.findByIdAndUpdate(req.params.id, { title, content }, { new: true });
                 res.json(post);
             } else {
-                res.status(404).json({ error: 'Post topilmadi' });
+                res.status(404).json({ error: 'Post topilmadi yoki xato so\'rov yubordingiz.' });
             }
         } catch (error) {
             res.status(500).json({ error: 'Postni yangilashda xatolik yuz berdi' });
@@ -90,8 +89,16 @@ class PostController {
     // Postni o'chirish
     async deletePost(req: Request, res: Response) {
         try {
-            const post: IPost | null = await Post.findByIdAndDelete(req.params.id);
-            if (post) {
+            let token: any = req.headers.token;
+            const id = JWT.VERIFY(token).id;
+            if (!id) {
+                return res.status(401).json({
+                    error: 'unknown'
+                });
+            }
+            let post: IPost | null = await Post.findById(req.params.id);
+            if (post && post.user == id) {
+                const post: IPost | null = await Post.findByIdAndDelete(req.params.id);
                 res.json({ message: 'Post o\'chirildi' });
             } else {
                 res.status(404).json({ error: 'Post topilmadi' });
@@ -101,5 +108,4 @@ class PostController {
         }
     }
 }
-
 export default new PostController();
